Added UIStringChangeDetector tests inside AppDelegate. Made some other slight adjustments based on our findings.
I think we'll move away from using NSAccessibility notifications for detecting uiString changes and just swizzle everything manually. That should hopefully be more robust.

Problems I saw with the NSAccessibilty Notifications:
For NSWindowTitle and NSWindowSubtitle 10 notifications are sent in a row. For NSCheckbox title/stringValue the notifications don't seem to be sent at all. For NSTextFieldCell the notification seems to be sent before the value actually changed. But for NSSearchFieldCell the notification is sent after the value is changed.

On 16. July we ammended this commit (accidentally)

Here are the notes:

- More views and stuff to test added to the IB file.
- Updated the NibAnnotation code to work with the new views.
- Added tests for the uiStringChange detection.
- Progress on subclass swizzling to intercept ui string setters across appkit. We have problems dispatch back to the original method reliably after a call has been intercepted. We have a hacky approach to solving that at the moment which would break for recursive functions and maybe other scenarios. I just had a new idea! Using C preprocessor macros to create and IMP factory which will get the right types through the macro and can then be used to inside the swizzling code to create interceptor imps which directly information about how to call the original method that was being intercepted. It sounds crazy but I think it'll work robustly! And it's the least complicated solution I can think of ...
- Probably other stuff we forgot.